refactor: move _remap_outputs function to rfd3.py and remove from rf3.py
